+++ perception/object_segmentation/ground_segmentation/include/ground_segmentation/ransac_ground_filter_nodelet.hpp
@@ -32,7 +32,7 @@
 #include <string>
 #include <vector>
 
-namespace ground_filter
+namespace ground_segmentation
 {
 struct PlaneBasis
 {
@@ -123,6 +123,6 @@ public:
   EIGEN_MAKE_ALIGNED_OPERATOR_NEW
   explicit RANSACGroundFilterComponent(const rclcpp::NodeOptions & options);
 };
-}  // namespace ground_filter
+}  // namespace ground_segmentation
 
 #endif  // POINTCLOUD_PREPROCESSOR__GROUND_FILTER__RANSAC_GROUND_FILTER_NODELET_HPP_
