+++ perception/object_segmentation/ground_segmentation/package.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0"?>
 <package format="3">
-  <name>ground_filter</name>
+  <name>ground_segmentation</name>
   <version>0.1.0</version>
-  <description>The ROS2 ground_filter package</description>
+  <description>The ROS2 ground_segmentation package</description>
 
   <maintainer email="abrahammonrroy@yahoo.com">amc-nu</maintainer>
   <maintainer email="yukihiro.saito@tier4.jp">Yukihiro Saito</maintainer>
