+++ perception/object_segmentation/ground_segmentation/src/ray_ground_filter_nodelet.cpp
@@ -29,14 +29,14 @@
  *  v1.0: amc-nu (abrahammonrroy@yahoo.com)
  */
 
-#include "ground_filter/ray_ground_filter_nodelet.hpp"
+#include "ground_segmentation/ray_ground_filter_nodelet.hpp"
 
 #include <pcl_ros/transforms.hpp>
 
 #include <string>
 #include <vector>
 
-namespace ground_filter
+namespace ground_segmentation
 {
 using pointcloud_preprocessor::get_param;
 
@@ -283,8 +283,8 @@ void RayGroundFilterComponent::ClassifyPointCloud(
 //   // Enable the dynamic reconfigure service
 //   has_service = true;
 //   srv_ = boost::make_shared<
-//     dynamic_reconfigure::Server<ground_filter::RayGroundFilterConfig> >(nh);
-//   dynamic_reconfigure::Server<ground_filter::RayGroundFilterConfig>::CallbackType f =
+//     dynamic_reconfigure::Server<ground_segmentation::RayGroundFilterConfig> >(nh);
+//   dynamic_reconfigure::Server<ground_segmentation::RayGroundFilterConfig>::CallbackType f =
 //     boost::bind(&RayGroundFilterComponent::config_callback, this, _1, _2);
 //   srv_->setCallback(f);
 //   return (true);
@@ -420,7 +420,7 @@ rcl_interfaces::msg::SetParametersResult RayGroundFilterComponent::paramCallback
   return result;
 }
 
-}  // namespace ground_filter
+}  // namespace ground_segmentation
 
 #include <rclcpp_components/register_node_macro.hpp>
-RCLCPP_COMPONENTS_REGISTER_NODE(ground_filter::RayGroundFilterComponent)
+RCLCPP_COMPONENTS_REGISTER_NODE(ground_segmentation::RayGroundFilterComponent)
